feat: useCarousel 훅으로 캐러셀 로직 분리 및 적용 (#292)
